refactor: Modal에서 calculateDisableCoupon를 함수로 분리
