+++ src/components/OrderList/Modal/Modal.tsx
@@ -8,67 +8,13 @@ import { CouponResponse } from "../../../types/Coupon";
 import * as S from "./Modal.styles";
 import CouponItem from "../Coupon/CouponItem";
 import CartItemCheck from "../../../types/CartItemCheck";
+import calculateDisableCoupon from "../../../utils/calculateDisableCoupon";
 
 interface ModalProps {
   isModalOpen: boolean;
   onClose: () => void;
   selectedCartItemList: CartItemCheck[];
   shippingFee: number;
-}
-
-function computeDisabled(
-  coupon: CouponResponse,
-  opts: {
-    couponIds: number[];
-    allProductPrice: number;
-    shippingFee: number;
-    selectedItems: { quantity: number }[];
-    now: Date;
-  }
-) {
-  const { couponIds, allProductPrice, shippingFee, selectedItems, now } = opts;
-  const isChecked = couponIds.includes(coupon.id);
-
-  if (!isChecked && couponIds.length >= 2) return true;
-
-  if (
-    (coupon.discountType === "fixed" ||
-      coupon.discountType === "freeShipping") &&
-    allProductPrice < coupon.minimumAmount
-  )
-    return true;
-
-  if (coupon.discountType === "freeShipping" && shippingFee === 0) return true;
-
-  if (
-    coupon.discountType === "buyXgetY" &&
-    !selectedItems.some(
-      (item) => item.quantity >= coupon.buyQuantity + coupon.getQuantity
-    )
-  )
-    return true;
-
-  if (coupon.discountType === "percentage" && coupon.availableTime) {
-    const [sh, sm] = coupon.availableTime.start.split(":").map(Number);
-    const [eh, em] = coupon.availableTime.end.split(":").map(Number);
-    const startTime = new Date(
-      now.getFullYear(),
-      now.getMonth(),
-      now.getDate(),
-      sh,
-      sm
-    );
-    const endTime = new Date(
-      now.getFullYear(),
-      now.getMonth(),
-      now.getDate(),
-      eh,
-      em
-    );
-    if (now < startTime || now > endTime) return true;
-  }
-
-  return false;
 }
 
 export default function Modal({
@@ -119,7 +65,7 @@ export default function Modal({
             </S.Info>
             <S.CouponList>
               {couponList.map((coupon) => {
-                const isDisabled = computeDisabled(coupon, {
+                const isDisabled = calculateDisableCoupon(coupon, {
                   couponIds,
                   allProductPrice,
                   shippingFee,
